// src/app/core/api/model/playerPayload.ts
/**
 * Sogexia Hiring Test API - OpenAPI 3.0
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Player used in POST request body to create a new player
 */
export interface PlayerPayload { 
    name?: string;
}