// src/app/core/api/model/score.ts
/**
 * Sogexia Hiring Test API - OpenAPI 3.0
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * One score for one player
 */
export interface Score { 
    playerId?: number;
    score?: number;
}